Combine helper files
